+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version 0.0.3 (2019-04-29)
 
-* Add the `smd_no_move!` macro
+* Add the `smd_borrowed!` macro
 * Add documentation
 * fix smd!() not compiling
 * fix unused import compiler warnings